+++ docs/lib/examples/Jumbotron.js
@@ -7,7 +7,7 @@ const Example = (props) => {
       <Jumbotron>
         <h1 className="display-3">Hello, world!</h1>
         <p className="lead">This is a simple hero unit, a simple Jumbotron-style component for calling extra attention to featured content or information.</p>
-        <hr className="m-y-2" />
+        <hr className="my-2" />
         <p>It uses utility classes for typgraphy and spacing to space content out within the larger container.</p>
         <p className="lead">
           <Button color="primary">Learn More</Button>