+++ docs/lib/examples/MediaAlignment.js
@@ -15,7 +15,7 @@ const Example = () => {
           Cras sit amet nibh libero, in gravida nulla. Nulla vel metus scelerisque ante sollicitudin commodo. Cras purus odio, vestibulum in vulputate at, tempus viverra turpis. Fusce condimentum nunc ac nisi vulputate fringilla. Donec lacinia congue felis in faucibus.
         </Media>
       </Media>
-      <Media className="m-t-1">
+      <Media className="mt-1">
         <Media left middle href="#">
           <Media object data-src="holder.js/64x64" alt="Generic placeholder image" />
         </Media>
@@ -26,7 +26,7 @@ const Example = () => {
           Cras sit amet nibh libero, in gravida nulla. Nulla vel metus scelerisque ante sollicitudin commodo. Cras purus odio, vestibulum in vulputate at, tempus viverra turpis. Fusce condimentum nunc ac nisi vulputate fringilla. Donec lacinia congue felis in faucibus.
         </Media>
       </Media>
-      <Media className="m-t-1">
+      <Media className="mt-1">
         <Media left bottom href="#">
           <Media object data-src="holder.js/64x64" alt="Generic placeholder image" />
         </Media>