+++ src/Col.js
@@ -1,5 +1,6 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 
 const colSizes = ['xs', 'sm', 'md', 'lg', 'xl'];
 const stringOrNumberProp = PropTypes.oneOfType([PropTypes.number, PropTypes.string]);
@@ -22,7 +23,8 @@ const propTypes = {
   md: columnProps,
   lg: columnProps,
   xl: columnProps,
-  className: PropTypes.node
+  className: PropTypes.node,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -32,6 +34,7 @@ const defaultProps = {
 const Col = (props) => {
   const {
     className,
+    cssModule,
     ...attributes
   } = props;
   const colClasses = [];
@@ -50,12 +53,12 @@ const Col = (props) => {
         colClass = `col-${colSize}-${columnProp.size}`;
       }
 
-      colClasses.push(classNames({
+      colClasses.push(mapToCssModules(classNames({
         [colClass]: columnProp.size,
         [`push-${colSize}-${columnProp.push}`]: columnProp.push,
         [`pull-${colSize}-${columnProp.pull}`]: columnProp.pull,
         [`offset-${colSize}-${columnProp.offset}`]: columnProp.offset
-      }));
+      })), cssModule);
     } else {
       if (columnProp === 'auto' || columnProp === true) {
         colClass = `col-${colSize}`;
@@ -67,10 +70,10 @@ const Col = (props) => {
     }
   });
 
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     colClasses
-  );
+  ), cssModule);
 
   return (
     <div {...attributes} className={classes} />