+++ src/FormFeedback.js
@@ -1,10 +1,12 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 
 const propTypes = {
   children: PropTypes.node,
   tag: PropTypes.string,
-  className: PropTypes.string,
+  className: PropTypes.any,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -14,14 +16,15 @@ const defaultProps = {
 const FormFeedback = (props) => {
   const {
     className,
+    cssModule,
     tag: Tag,
     ...attributes,
   } = props;
 
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     'form-control-feedback'
-  );
+  ), cssModule);
 
   return (
     <Tag {...attributes} className={classes} />