+++ src/FormText.js
@@ -1,12 +1,14 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 
 const propTypes = {
   children: PropTypes.node,
   inline: PropTypes.bool,
   tag: PropTypes.string,
   color: PropTypes.string,
-  className: PropTypes.string,
+  className: PropTypes.any,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -16,17 +18,18 @@ const defaultProps = {
 const FormText = (props) => {
   const {
     className,
+    cssModule,
     inline,
     color,
     tag: Tag,
     ...attributes,
   } = props;
 
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     !inline ? 'form-text' : false,
     color ? `text-${color}` : false
-  );
+  ), cssModule);
 
   return (
     <Tag {...attributes} className={classes} />