+++ src/Jumbotron.js
@@ -1,10 +1,12 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 
 const propTypes = {
   tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
   fluid: PropTypes.bool,
-  className: PropTypes.any
+  className: PropTypes.any,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -14,16 +16,17 @@ const defaultProps = {
 const Jumbotron = (props) => {
   const {
     className,
+    cssModule,
     tag: Tag,
     fluid,
     ...attributes
   } = props;
 
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     'jumbotron',
     fluid ? 'jumbotron-fluid' : false
-  );
+  ), cssModule);
 
   return (
     <Tag {...attributes} className={classes} />