+++ src/NavDropdown.js
@@ -1,11 +1,13 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 import Dropdown from './Dropdown';
 
 const propTypes = {
   children: PropTypes.node,
   tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
-  className: PropTypes.any
+  className: PropTypes.any,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -15,14 +17,15 @@ const defaultProps = {
 const NavDropdown = (props) => {
   const {
     className,
+    cssModule,
     tag: Tag,
     ...attributes
   } = props;
 
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     'nav-item'
-  );
+  ), cssModule);
 
   return (
     <Dropdown {...attributes} tag={Tag} className={classes} />