+++ src/Progress.js
@@ -1,5 +1,6 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 import toNumber from 'lodash.tonumber';
 
 const propTypes = {
@@ -15,7 +16,8 @@ const propTypes = {
   animated: PropTypes.bool,
   striped: PropTypes.bool,
   color: PropTypes.string,
-  className: PropTypes.any
+  className: PropTypes.any,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -27,6 +29,7 @@ const defaultProps = {
 const Progress = (props) => {
   const {
     className,
+    cssModule,
     value,
     max,
     animated,
@@ -38,23 +41,23 @@ const Progress = (props) => {
 
   const percent = ((toNumber(value) / toNumber(max)) * 100);
 
-  const nonProgressClasses = classNames(
+  const nonProgressClasses = mapToCssModules(classNames(
     className,
     'progress',
     animated ? 'progress-animated' : null
-  );
+  ), cssModule);
 
-  const progressClasses = classNames(
+  const progressClasses = mapToCssModules(classNames(
     nonProgressClasses,
     color ? `progress-${color}` : null,
     striped || animated ? 'progress-striped' : null
-  );
+  ), cssModule);
 
-  const fallbackClasses = classNames(
+  const fallbackClasses = mapToCssModules(classNames(
     'progress-bar',
     color ? `progress-${color}` : null,
     striped || animated ? 'progress-bar-striped' : null
-  );
+  ), cssModule);
 
   const fallbackFill = (
     <span