+++ src/Row.js
@@ -1,9 +1,11 @@
 import React, { PropTypes } from 'react';
 import classNames from 'classnames';
+import { mapToCssModules } from './utils';
 
 const propTypes = {
   tag: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
-  className: PropTypes.any
+  className: PropTypes.any,
+  cssModule: PropTypes.object,
 };
 
 const defaultProps = {
@@ -13,14 +15,15 @@ const defaultProps = {
 const Row = (props) => {
   const {
     className,
+    cssModule,
     tag: Tag,
     ...attributes
   } = props;
 
-  const classes = classNames(
+  const classes = mapToCssModules(classNames(
     className,
     'row'
-  );
+  ), cssModule);
 
   return (
     <Tag {...attributes} className={classes} />